+++ app/src/main/res/layout/activity_main.xml
@@ -25,6 +25,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
+        app:labelVisibilityMode="labeled"
         app:menu="@menu/bottom_navigation"
         android:background="?android:attr/windowBackground"/>
