--- app/src/main/res/layout/fragment_add.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent" android:layout_height="match_parent"
-    android:background="#C4E1E4">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Add ingredients by scanning or manually entering them"
-        android:textSize="30sp"
-        android:layout_centerInParent="true"/>
-
-</RelativeLayout>
